fashion-ai-pipeline/requirements.txt
torch
torchvision
opencv-python
supervision
ultralytics
Pillow
requests
python-dotenv
psycopg2-binary
supabase
